Support new links in cli
