+++ apps/builder/app/services/builder-auth.server.ts
@@ -9,8 +9,8 @@ import {
 import {
   getAuthorizationServerOrigin,
   getRequestOrigin,
-  parseBuilderUrl,
 } from "~/shared/router-utils/origins";
+import { parseBuilderUrl } from "@webstudio-is/http-client";
 import { createDebug } from "~/shared/debug";
 import { readAccessToken } from "./token.server";
 import { isUserAuthorizedForProject } from "./builder-access.server";
