+++ packages/cli/src/commands/link.ts
@@ -2,6 +2,7 @@ import { cwd, exit } from "node:process";
 import { join } from "node:path";
 import { readFile, writeFile } from "node:fs/promises";
 import { cancel, isCancel, log, text } from "@clack/prompts";
+import { parseBuilderUrl } from "@webstudio-is/http-client";
 import {
   GLOBAL_CONFIG_FILE,
   LOCAL_CONFIG_FILE,
@@ -17,19 +18,27 @@ import type {
 
 const parseShareLink = (value: string) => {
   const url = new URL(value);
-  const origin = url.origin;
+
   const token = url.searchParams.get("authToken");
-  // @todo parse with URLPattern once landed in node
-  const segments = url.pathname.split("/").slice(1);
-  if (segments.length !== 2 || segments[0] !== "builder") {
-    throw Error("Segments not matching");
+
+  // eslint-disable-next-line prefer-const
+  let { projectId, sourceOrigin } = parseBuilderUrl(url.href);
+
+  if (projectId === undefined) {
+    // Support deprecated links until the end of 2024
+    const segments = url.pathname.split("/").slice(1);
+    if (segments.length !== 2 || segments[0] !== "builder") {
+      throw Error("Segments not matching");
+    }
+    projectId = segments[1];
   }
-  const [_builder, projectId] = segments;
+
   if (token == null) {
     throw Error("Token is missing");
   }
+
   return {
-    origin,
+    origin: sourceOrigin,
     projectId,
     token,
   };
