+++ packages/http-client/jest.config.js
@@ -1,0 +1,1 @@
+export { default } from "@webstudio-is/jest-config";
