+++ packages/http-client/package.json
@@ -8,12 +8,15 @@
   "scripts": {
     "build": "rm -rf lib && esbuild src/index.ts --outdir=lib --bundle --format=esm --packages=external",
     "dts": "tsc --project tsconfig.dts.json",
-    "typecheck": "tsc"
+    "typecheck": "tsc",
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest"
   },
   "dependencies": {
     "@webstudio-is/sdk": "workspace:*"
   },
   "devDependencies": {
+    "@jest/globals": "^29.7.0",
+    "@webstudio-is/jest-config": "workspace:*",
     "@webstudio-is/tsconfig": "workspace:*",
     "typescript": "5.5.2"
   },
